fennecs 0.4.0-beta support, and moved Capabilities class into Capabilities.cs, and changed the --ranking option to a RANK_RESULTS define
