+++ source/Ecs.CSharp.Benchmark/CreateEntityWithThreeComponents/Fennecs.cs
@@ -11,18 +11,16 @@ namespace Ecs.CSharp.Benchmark
         [Context] private readonly FennecsBaseContext _fennecs;
         
         [BenchmarkCategory(Categories.Fennecs)]
-        [Benchmark]
+        [Benchmark(Description = "fennecs")]
         public void Fennecs()
         {
             World world = _fennecs.World;
 
-            for (int i = 0; i < EntityCount; ++i)
-            {
-                world.Spawn()
-                    .Add<Component1>()
-                    .Add<Component2>()
-                    .Add<Component3>();
-            }
+            world.Entity()
+                .Add(new Component1())
+                .Add(new Component2())
+                .Add(new Component3())
+                .Spawn(EntityCount);
         }
     }
 }
