+++ source/Ecs.CSharp.Benchmark/CreateEntityWithTwoComponents/Fennecs.cs
@@ -10,16 +10,15 @@ namespace Ecs.CSharp.Benchmark
         [Context] private readonly FennecsBaseContext _fennecs;
         
         [BenchmarkCategory(Categories.Fennecs)]
-        [Benchmark]
+        [Benchmark(Description = "fennecs")]
         public void Fennecs()
         {
             World world = _fennecs.World;
 
-            for (int i = 0; i < EntityCount; ++i)
-            {
-                world.Spawn().
-                    Add<Component1>().Add<Component2>();
-            }
+            world.Entity()
+                .Add(new Component1())
+                .Add(new Component2())
+                .Spawn(EntityCount);
         }
     }
 }
