+++ source/Ecs.CSharp.Benchmark/Program.cs
@@ -1,10 +1,7 @@
 ﻿#pragma warning disable CA1852 // Seal internal types
 
-using System.Collections.Generic;
 using System.Globalization;
-using System.Linq;
 using BenchmarkDotNet.Configs;
-using BenchmarkDotNet.Order;
 using BenchmarkDotNet.Running;
 using Ecs.CSharp.Benchmark;
 
@@ -34,17 +31,12 @@ IConfig configuration = DefaultConfig.Instance
     .WithOptions(ConfigOptions.DisableOptimizationsValidator)
     .WithCapabilityExclusions();
 
+#if RANK_RESULTS
+    configuration = configuration.WithOrderer(new BenchmarkDotNet.Order.DefaultOrderer(BenchmarkDotNet.Order.SummaryOrderPolicy.FastestToSlowest));
+#endif
+
 if (args.Length > 0)
 {
-    // There's no orderer commandline arg, so let's pretend there is one.
-    if (args.Contains("--ranking"))
-    {
-        List<string> argList = args.ToList();
-        argList.Remove("--ranking");
-        args = argList.ToArray();
-        configuration = configuration.WithOrderer(new DefaultOrderer(SummaryOrderPolicy.FastestToSlowest));
-    }
-
     benchmark.Run(args, configuration);
 }
 else
